Rename Homeserver models to PubkyApp
