--- src/models/homeserver/mod.rs
@@ -1,13 +1,0 @@
-/// Raw schemas as stored on homeserver.
-mod bookmark;
-mod follow;
-mod post;
-mod tag;
-pub mod traits;
-mod user;
-
-pub use bookmark::HomeserverBookmark;
-pub use follow::HomeserverFollow;
-pub use post::{HomeserverPost, PostKind};
-pub use tag::HomeserverTag;
-pub use user::{HomeserverUser, UserLink};
