+++ src/models/mod.rs
@@ -1,6 +1,6 @@
-pub mod homeserver;
 pub mod info;
 pub mod post;
+pub mod pubky_app;
 pub mod tag;
 pub mod traits;
 pub mod user;
