+++ src/models/pubky_app/tag.rs
@@ -12,24 +12,24 @@ use serde::{Deserialize, Serialize};
 ///
 /// Where tag_id is z-base32(Sha256("{uri_tagged}:{")))[:8]
 #[derive(Serialize, Deserialize, Default)]
-pub struct HomeserverTag {
+pub struct PubkyAppTag {
     pub uri: String,
     pub label: String,
     pub created_at: i64,
 }
 
-impl GenerateId for HomeserverTag {
+impl GenerateId for PubkyAppTag {
     /// Tag ID is created based on the hash of the URI tagged and the label used
     fn get_id_data(&self) -> String {
         format!("{}:{}", self.uri, self.label)
     }
 }
 
-impl HomeserverTag {
-    /// Creates a unique identifier (tag ID) for the `HomeserverTag` instance.
+impl PubkyAppTag {
+    /// Creates a unique identifier (tag ID) for the `PubkyAppTag` instance.
     ///
     /// The tag ID is generated by:
-    /// 1. Concatenating the `uri` and `label` fields of the `HomeserverTag` with a colon (`:`) separator.
+    /// 1. Concatenating the `uri` and `label` fields of the `PubkyAppTag` with a colon (`:`) separator.
     /// 2. Hashing the concatenated string using the `blake3` hashing algorithm.
     /// 3. Taking the first half of the bytes from the resulting `blake3` hash.
     /// 4. Encoding those bytes using the Z-base32 alphabet (Base32 variant).
@@ -60,7 +60,7 @@ impl HomeserverTag {
 
 #[test]
 fn testcreate_id() {
-    let tag = HomeserverTag {
+    let tag = PubkyAppTag {
         uri: "user_id/pub/pubky.app/posts/post_id".to_string(),
         created_at: 1627849723,
         label: "cool".to_string(),
