+++ src/models/pubky_app/traits.rs
@@ -8,7 +8,7 @@ pub trait GenerateId {
     /// Creates a unique identifier for bookmarks and tag homeserver paths instance.
     ///
     /// The ID is generated by:
-    /// 1. Concatenating the `uri` and `label` fields of the `HomeserverTag` with a colon (`:`) separator.
+    /// 1. Concatenating the `uri` and `label` fields of the `PubkyAppTag` with a colon (`:`) separator.
     /// 2. Hashing the concatenated string using the `blake3` hashing algorithm.
     /// 3. Taking the first half of the bytes from the resulting `blake3` hash.
     /// 4. Encoding those bytes using the Z-base32 alphabet (Base32 variant).
