+++ src/routes/v0/user/details.rs
@@ -1,4 +1,4 @@
-use crate::models::homeserver::UserLink;
+use crate::models::pubky_app::UserLink;
 use crate::models::user::UserDetails;
 use crate::routes::v0::endpoints::USER_DETAILS_ROUTE;
 use crate::{Error, Result};
